requirements.txt
browserstack-local
jsonmerge
multiprocess
paver
selenium==4.6.0
psutil
pytest==7.1
pytest-variables
pytest-selenium
Appium-Python-Client
browserstack-sdk